feat: new invoices logic (GTS-2228)

+++ src/modules/Client/Invoicing/Application/UseCase/CancelOrderInvoice.php
@@ -2,9 +2,13 @@
 
 namespace Module\Client\Invoicing\Application\UseCase;
 
+use Module\Client\Invoicing\Application\Exception\InvalidOrderStatusToCancelInvoiceException;
+use Module\Client\Invoicing\Domain\Invoice\Exception\InvalidOrderStatusToCancelInvoice;
 use Module\Client\Invoicing\Domain\Invoice\Repository\InvoiceRepositoryInterface;
+use Module\Client\Invoicing\Domain\Order\Order;
 use Module\Client\Invoicing\Domain\Order\Repository\OrderRepositoryInterface;
 use Module\Client\Invoicing\Domain\Order\ValueObject\OrderId;
+use Module\Shared\Exception\ApplicationException;
 use Sdk\Module\Contracts\UseCase\UseCaseInterface;
 
 class CancelOrderInvoice implements UseCaseInterface
@@ -12,17 +16,31 @@ class CancelOrderInvoice implements UseCaseInterface
     public function __construct(
         private readonly OrderRepositoryInterface $orderRepository,
         private readonly InvoiceRepositoryInterface $invoiceRepository,
-    ) {
-    }
+    ) {}
 
     public function execute(int $orderId): void
     {
         $order = $this->orderRepository->findOrFail(new OrderId($orderId));
 
-        $order->ensureInvoiceCanBeCancelled();
+        try {
+            $order->ensureInvoiceCanBeCancelled();
+        } catch (InvalidOrderStatusToCancelInvoice $e) {
+            throw new InvalidOrderStatusToCancelInvoiceException($e);
+        }
 
-        $invoice = $this->invoiceRepository->find($order->invoiceId());
+        $invoice = $this->invoiceRepository->findByOrderId($order->id());
+        if ($invoice === null) {
+            throw new ApplicationException('Not found invoice for order', 404);
+        }
         $invoice->delete();
         $this->invoiceRepository->store($invoice);
+
+        $this->updateOrderStatus($order);
+    }
+
+    public function updateOrderStatus(Order $order): void
+    {
+        $order->toInProgress();
+        $this->orderRepository->store($order);
     }
 }
